+++ examplepatches/franticg33k/morphe-patches/patches/src/main/kotlin/app/franticg33k/patches/nativecamera/shared/Constants.kt
@@ -1,6 +1,7 @@
 package app.franticg33k.patches.nativecamera.shared
 
 import app.morphe.patcher.patch.ApkFileType
+import app.morphe.patcher.patch.AppTarget
 import app.morphe.patcher.patch.Compatibility
 
 object Constants {
@@ -9,5 +10,12 @@ object Constants {
         packageName = "com.rawcam.app",
         apkFileType = ApkFileType.APK,
         appIconColor = 0xd0bcff,
+        targets = listOf(
+            AppTarget(
+                version = "1.1.7",
+                isExperimental = false,
+                minSdk = null,
+            ),
+        ),
     )
 }